Separeted services from database
