+++ services/assistant/app/database/__init__.py
@@ -1,5 +1,4 @@
 from database.orm import AbstractRepository
 from database.database import get_db_factory
-from database.services import ServiceFactory
 
-__all__ = ["AbstractRepository", "get_db_factory", "ServiceFactory"]
+__all__ = ["AbstractRepository", "get_db_factory"]
