+++ services/assistant/app/services/__init__.py
@@ -1,0 +1,5 @@
+from services.service_factory import ServiceFactory
+from services.user_service import UserService
+from services.booking_service import BookingService
+
+__all__ = ["ServiceFactory", "UserService", "BookingService"]
